refactor: move date constants out of type file
