+++ src/common/date.ts
@@ -1,0 +1,12 @@
+export enum DayOfWeek {
+  Sunday = 0,
+  Monday = 1,
+  Tuesday = 2,
+  Wednesday = 3,
+  Thursday = 4,
+  Friday = 5,
+  Saturday = 6,
+}
+
+export const DEFAULT_TIMEZONE = 'Etc/UTC';
+export const DEFAULT_WEEK_START = DayOfWeek.Monday;
