+++ src/cron/dailyDigest.ts
@@ -3,6 +3,7 @@ import {
   schedulePersonalizedDigestSubscriptions,
   digestPreferredHourOffset,
   notifyGeneratePersonalizedDigest,
+  DEFAULT_TIMEZONE,
 } from '../common';
 import {
   User,
@@ -12,7 +13,6 @@ import {
 } from '../entity';
 import { Cron } from './cron';
 import { isWeekend, addHours, startOfHour, subDays } from 'date-fns';
-import { DEFAULT_TIMEZONE } from '../types';
 
 const sendType = UserPersonalizedDigestSendType.workdays;
 const digestTypes = [
